pyproject.toml
[tool.poetry]
name = "secure_t"
version = "0.1.0"
description = ""
authors = ["painassasin <painassasin@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = {version = "^0.18.3", extras = ["standard"]}
asyncpg = "^0.26.0"
SQLAlchemy = {version = "^1.4.41", extras = ["asyncio"]}
alembic = "^1.8.1"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
PyJWT = "^2.5.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
